Rename `types.rs` to `geo.rs`
